Don't change the monasca-transform user if in use (bsc#1092182)
It was found in update testing that the ansible script was attempting
to change the monasca-persister user but there are processes running
as the monasca-persister user, which caused a failure. Monasca-transform
has similar code, so we are fixing it here as well.

This change catches the error from the attempt to change the existing
user and if the failure is due to existing processes under the
monasca-transform user it continues on. The advantage of this approach
is that if the monasca-transform processes are stopped before the
install.yml is run then the change can be made.

Change-Id: I8910e90eb4e71a9d47d0100260409303cbc07d72